+++ packages/atomic-router/src/chain.ptorocol.ts
@@ -1,8 +1,8 @@
-import { type EventAsReturnType } from 'effector';
-import { type RouteParamsAndQuery } from 'atomic-router';
+import type { Event, EventCallable } from 'effector';
+import type { RouteParamsAndQuery } from 'atomic-router';
 
 export type ChainProtocol<RouteParams extends Record<string, any>> = {
-  beforeOpen: EventAsReturnType<RouteParamsAndQuery<RouteParams>>;
-  openOn: EventAsReturnType<any>;
-  cancelOn: EventAsReturnType<any>;
+  beforeOpen: EventCallable<RouteParamsAndQuery<RouteParams>>;
+  openOn: Event<any>;
+  cancelOn: Event<any>;
 };