+++ packages/core/src/libs/lohyphen/index.ts
@@ -7,3 +7,5 @@ export { isEmpty, isNotEmpty } from './is_empty';
 export { isEqual } from './is_equal';
 export { divide } from './divide';
 export { get } from './field';
+export { uniq } from './uniq';
+export { Mutex } from './mutex';