+++ packages/core/src/libs/patronus/abortable.ts
@@ -77,7 +77,11 @@ export function abortable<P = void, D = void, F = Error>(config: {
 
     return result;
   });
-  const $calls = createStore<CurrentCall[]>([], { serialize: 'ignore' });
+  const $calls = createStore<CurrentCall[]>([], {
+    serialize: 'ignore',
+    name: 'ff.$calls',
+    sid: 'ff.$calls',
+  });
   const callsApi = createApi($calls, {
     add(calls, def: CurrentCall) {
       return [...calls, def];
@@ -88,13 +92,17 @@ export function abortable<P = void, D = void, F = Error>(config: {
   });
 
   if (abort?.signal) {
-    const abortTrigger = sample({ clock: abort.signal });
-
-    $calls.watch(abortTrigger, (calls) => {
+    const cancelAllFx = createEffect((calls: CurrentCall[]) => {
       calls.forEach((c) => {
         c.context?.runAborters?.();
         c.reject(abortError());
       });
+    });
+
+    sample({
+      clock: abort.signal,
+      source: $calls,
+      target: cancelAllFx,
     });
   }