+++ packages/core/src/libs/patronus/readonly.ts
@@ -6,5 +6,5 @@ export function readonly<T>(event: Event<T>): Event<T>;
 export function readonly<T>(
   storeOrEvent: Store<T> | Event<T>
 ): Store<T> | Event<T> {
-  return storeOrEvent.map((v) => v);
+  return storeOrEvent.map((v) => v, { skipVoid: false });
 }