+++ packages/core/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @farfetched/core
 
+## 0.11.1
+
+### Patch Changes
+
+- 730ef73: Do not flick `.$status` during cancellation
+
+## 0.11.0
+
+### Minor Changes
+
+- c925412: Introduce Barrier API
+- 67d7288: Hide `isAbortError` from public API
+- ffdfffa: Adopt for Effector 23
+
+### Patch Changes
+
+- ef2d3a5: Support `enabled` in `keepFresh`
+- ec2ab95: Do not throw nonserializable error in case of invalid URL
+
 ## 0.10.6
 
 ### Patch Changes